Deduplicate academic calendar and budget functions
